r/aws_ec2_secondary_subnet: Move finders etc.
