+++ internal/service/ec2/consts.go
@@ -151,16 +151,6 @@ func vpnTunnelOptionsIKEVersion_Values() []string {
 	}
 }
 
-// See https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_SecondarySubnet.html.
-// Note: Secondary subnet states are available as AWS SDK enum types (awstypes.SecondarySubnetState*)
-// but we define constants here for consistency and potential future use.
-const (
-	SecondarySubnetStateCreateComplete   = "create-complete"
-	SecondarySubnetStateCreateInProgress = "create-in-progress"
-	SecondarySubnetStateDeleteComplete   = "delete-complete"
-	SecondarySubnetStateDeleteInProgress = "delete-in-progress"
-)
-
 const (
 	vpnTunnelCloudWatchLogOutputFormatJSON = names.AttrJSON
 	vpnTunnelCloudWatchLogOutputFormatText = "text"
