+++ internal/service/ec2/status.go
@@ -1900,3 +1900,19 @@ func statusSecondaryNetwork(ctx context.Context, conn *ec2.Client, id string) sd
 		return output, string(output.State), nil
 	}
 }
+
+func statusSecondarySubnet(ctx context.Context, conn *ec2.Client, id string) sdkretry.StateRefreshFunc {
+	return func() (any, string, error) {
+		output, err := findSecondarySubnetByID(ctx, conn, id)
+
+		if retry.NotFound(err) {
+			return nil, "", nil
+		}
+
+		if err != nil {
+			return nil, "", err
+		}
+
+		return output, string(output.State), nil
+	}
+}
